refactor ansel-remote-import:
Start making it a more general purpose cli utility for Ansel.
For now, this just combines the functionality of the
recursive_import script. This one script now acts either locally or
remotely. Still planning on bringing in functionality from ansel.php.
Yes, classes are all in single file for now. Need to think through
if we can somehow package this in a small bundle to install on a
client machine without the need for an entire horde install.