Experiment with using 2 layers for image view.
Commiting this to get it into the repo, but plan on reverting.
Experiment with using 2 vector layers on image view so the
image thumbnails can be turned on/off while leaving the "place"
marker showing the location of the current image. Useful for
galleries with lots of images in close proximity.

It works, but the problem is that one of the following happens;
either the "place marker" is below the image thumbnails and is
potentially hidden by them, or the image thumbnails are behind
the place marker, but receive no events (so no hover or click).
This is a limitation of OL, as only the top most overlay receives
events.