Don't center the map after each marker is added.
Only need to center the map once, and do it when the
current image's marker is added. Still to decide: Should
the current image be in the center, with the map zoomed to
a reasonable zoomLevel - or should we zoomToFit the map, which
would ensure all geotagged images from the current gallery would
be shown on the map initially, but would not have the current image's
marker centered.