Consistent usage of 'smartmobile' instead of 'mobile'
The registry view name is 'smartmobile' (reasoning: mobile by itself is not
descriptive enough, and is too close in meaning to the minimal view -
which in the past has also been referred to as the mobile view).

Whether right or wrong, at least be consistent in usage.  CSS theming
already uses the 'smartmobile' nomenclature.